+++ macos/Test-Dependencies.ps1
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2024-2025 Friedrich von Never <friedrich@fornever.me>
+#
+# SPDX-License-Identifier: BSL-1.0
+
 param (
     [Parameter(Mandatory = $true)]
     [string] $DotNetArch,
@@ -55,7 +59,7 @@ if ($GenerateGold) {
     $goldContent = Get-Content -Raw $GoldFile
     $tempContent = Get-Content -Raw $ResultFile
     if ($goldContent -ne $tempContent) {
-        Write-Output "Current contents are following:`n"
+        Write-Output "Expected contents are following:`n"
         Write-Output $tempContent
         throw "File contents are not equal: $GoldFile and $ResultFile"
     }